Work around MSVC's old OpenMP version
MSVC only comes with OpenMP 2.0 which doesn't support unsigned loop counters or atomic read.
This commit replaces the loop counters used with openmp loops with their signed version and adds a comment to (hopefully) prevent people changing it back later.
A preprocessor ifdef is used to use `omp critical` in place of `omp atomic read` which is a bit heavy handed but will ensure the read is indeed atomic.

+++ src/TreeSupport.cpp
@@ -159,7 +159,8 @@ void TreeSupport::collisionAreas(const SliceDataStorage& storage, std::vector<st
     constexpr bool include_helper_parts = false;
     size_t completed = 0; //To track progress in a multi-threaded environment.
 #pragma omp parallel for shared(model_collision, storage) schedule(dynamic)
-    for (size_t radius_sample = 0; radius_sample < model_collision.size(); radius_sample++)
+    // Use a signed type for the loop counter so MSVC compiles
+    for (int radius_sample = 0; radius_sample < model_collision.size(); radius_sample++)
     {
         const coord_t radius = radius_sample * radius_sample_resolution;
         for (size_t layer_nr = 0; layer_nr < storage.support.supportLayers.size(); layer_nr++)
@@ -197,7 +198,8 @@ void TreeSupport::drawCircles(SliceDataStorage& storage, const std::vector<std::
     const coord_t line_width = storage.getSettingInMicrons("support_line_width");
     size_t completed = 0; //To track progress in a multi-threaded environment.
 #pragma omp parallel for shared(storage, contact_nodes)
-    for (size_t layer_nr = 0; layer_nr < contact_nodes.size(); layer_nr++)
+    // Use a signed type for the loop counter so MSVC compiles
+    for (int layer_nr = 0; layer_nr < contact_nodes.size(); layer_nr++)
     {
         Polygons support_layer;
         Polygons& roof_layer = storage.support.supportLayers[layer_nr].support_roof;
@@ -599,7 +601,7 @@ void TreeSupport::propagateCollisionAreas(const SliceDataStorage& storage, const
     const coord_t maximum_move_distance = angle < 90 ? (coord_t)(tan(angle) * layer_height) : std::numeric_limits<coord_t>::max();
     size_t completed = 0; //To track progress in a multi-threaded environment.
 #pragma omp parallel for shared(model_avoidance) schedule(dynamic)
-    for (size_t radius_sample = 0; radius_sample < model_avoidance.size(); radius_sample++)
+    for (int radius_sample = 0; radius_sample < model_avoidance.size(); radius_sample++)
     {
         model_avoidance[radius_sample].push_back(model_collision[radius_sample][0]);
         for (size_t layer_nr = 1; layer_nr < storage.support.supportLayers.size(); layer_nr++)
